Extract LLMNode module
